# Improve HyperKZG commitment performance with rayon parallelization