--- a/src/provider/hyperkzg.rs
+++ b/src/provider/hyperkzg.rs
@@ -486,8 +486,8 @@ where
     let h = <E::GE as DlogGroup>::group(&ck.h);
 
     E::GE::batch_vartime_multiscalar_mul(v, &ck.ck[..max])
-      .iter()
-      .zip(r.iter())
+      .par_iter()
+      .zip(r.par_iter())
       .map(|(commit, r_i)| Commitment {
         comm: *commit + (h * r_i),
       })
@@ -823,7 +823,7 @@ where
     // Compute commitments in parallel
     let r = vec![E::Scalar::ZERO; ell - 1];
     let com: Vec<G1Affine<E>> = E::CE::batch_commit(ck, &polys[1..], r.as_slice())
-      .iter()
+      .par_iter()
       .map(|i| i.comm.affine())
       .collect();